refactor(graph): add theme to graph
